Fix package.json: revert changes from CoalitionsUS mistakenly added to PR #57.
I submitted PR 57 from branch google-oauth and then several days later
made a change on that branch that should not have been included in the PR.
That change modified package.json so that the package was named mongoose-auth-coalitions,
and declared the source repository to be https://github.com/coalitions/mongoose-auth/

This change should revert package.json back to the correct state.
Note that I have taken the liberty to bump the version number to 0.0.12,
but otherwise the file is identical to the last revision by bnoguchi,
from commit 31578a8693b16d6d7cd1367816eb00ba045f84d3.